clm4_0_01
Add ability to handle glacier multiple elevation classes (glc_mec) in clm, so that we
can interact with the active glacier component (glc). Adds glacier elevation classes
to the surface datasets and requires they be read in when glacier multiple elevation
classes are active.  New namelist options for glc_mec include glc_smb and glc_dyntopo.
At build-time the number of glc_mec classes is set (can be 0, 1, 3, 5, or 10). The
model also interacts with the mask of valid glacier points that the active glacier
model determined (input with the fglcmask file), and set by glc_grid (which can be
gland5,gland10, or gland20 for 5-20km resolution over Greenland). glc_grid is set at
build time, but should be moved to the build-namelist.