clm4_0_16
Fix bug 1230, that caused problems with runoff to the ocean when running fully coupled. The global integrals of runoff fields
was the same in the coupler -- but the values where roundoff different. This caused problems both in testing for bit-for-bit with
the previous version and with restarts. The problem was that in the downscaling changes made in clm4_0_15 the areal scaling factor
asca needed to be copied from adomain into ldomain is no downscaling is taking place.