clm4_0_23
Add in new datasets and use-cases for glc_mec to support glc_nec=10 for 1850, 2000,
1850-2000, and 1850-2100 for all 4 rcp's. Standardize naming convention for use-cases.
Use scripts branch that has new compsets in it that access the new use-cases. Make sure
ncdpio is used for all I/O. Work with PTCLM a bit, and PTCLM testing. Change precedence
for build-namelist so that use-case is lower after user_nl_clm.