Create sub-cell integration utility function.
Given a field defined on the unit element, [-1,1]x[-1,1],
and a function defined by sample values and integration
weights at a certain number of Gauss-Lobatto-Legendre
points, divide the square up into sub-squares so that
there are now a bunch of sub-cells.

Integrate the function defined by the sample values
and weights over each of these sub-cells and return
the integrated values as a matrix.

Efficiency is obtained by computing and caching the
appropriate integration matrix the first time the
function is called.



git-svn-id: file:///Users/mataylo/repos/homme/trunk@2617 b670e762-8b1b-0410-9e79-e6a6fc92c689

Former-commit-id: 22231510efed047ff2911f51528dd59c09fb7372