Initial implementation of an arbitrary Lagrangian-Eularian tracer advection scheme.
The new use_semi_lagrange_transport flag will trigger the use of a new
tracer advection scheme. Presently the scheme simply does a backward
ALE step to determine the new tracer value and does a crude clipping
to preserve variation. There is no attempt at conservation at this
time. This crude algorithm will be expanded in the future to include
a conservation restriction. The commit to the code base at this time
is to facilitate code development and testing among the programming
team and is not meant as an option for tracer advection.



git-svn-id: file:///Users/mataylo/repos/homme/trunk@3601 b670e762-8b1b-0410-9e79-e6a6fc92c689

Former-commit-id: cb8961ba44924caf12aef47f9888f4710232a2f3