clm4_5_73
Add phys=clm5_0 as an option to the build. Currently, nothing is different in the code, this just
adds the capability to do clm5_0 as a seperate configuration. The one thing that is different between
clm4_5 and clm5_0 is the setting of urban_hac.

Fix several issues needed for CAM: problem in DryDeposition (reoccurance of bug 1883, that was fixed and then
unfixed in clm4_5_48), fix for internal compiler errors. CAM has been using the branch version of this
since: cam5_3_29.

Work on updates for PTCLM. Add some new sites for Rosie and Jinyun. Correct the call to mkmapdata.sh.
Have CLM1PT forcing directory to use DIN_LOC_ROOT_CLMFORC so you can point it to a location seperate
from DIN_LOC_ROOT. Add a new support script to PTCLM to submit a list of sites to batch: PTCLMsublist.
Also allow release_tags in version find. Get buildtools to work on edison/hopper.

Get tools to work on hopper and edison, and update mapping to use ESMF6.3.0.

Fix various bugs: internal compiler error on janus, trigger an error if user_datm.streams.txt file is
                  readonly (rather than hang). Use DIN_LOC_ROOT_CLMFORC for CLM1PT.

Requirements for tag: Fix bug 1883 and 1985 for Cheryl and CAM, fix PTCLM, add clm5_0, tools on hopper/edison