Modifying performance instrumentation in spectral element atmosphere dynamical core.
Fine tuning of profiling in CAM SE to eliminate events that occur at
too low a level (and thus are perturbed by the frequent timer calls)
and to add timers that provide call site information for functions that
are called from many different locations. The focus here is primarily
on the boundary exchange operator, in all of its forms.