CESMSCRATCHROOT on Mira (BFB)
This commit addresses removes extraneous directoris added to CESMSCRATCHROOT on Mira.

On all other platforms, the CESM scratch directory has the form:

/path/to/scratchsystem/user/casename
And then within this directory, there are various subdirectorires such as bld (where the executable is build and stored), run, where the data is written while the model is running.

On Mira, CESMSCRATCHROOT is set to
/projects/$PROJECT/$CCSMUSER/$CASE/scratch/exe

The addition of "scratch/exe" is pointless.

In For reference, the archive directory on Mira is set to:
/projects/$PROJECT/$CCSMUSER/archive/$CASE