Updating Mira makefile to pass ACME tests
This change updates machine-specific makefile Depends.mira and
reverts commit #d00074c, which was intended to fix a build problem
in the case '-compset ICRUCLM45 -res f09_f09 -mach mira'. In this
case, threading is disabled by default, but shr_reprosum_mod.F90
is compiled with qsmp flag, which leads to undefined references.
That commit led to run-time errors (signal 6 / SIGABRT) in ACME
tests. With this commit, ACME tests pass on Mira and the I case
above also builds and runs.