fixing support for pgicuda on Titan in config_compilers.xml
fixing support for pgicuda on Titan in config_compilers.xml

With the addition of Intel as an option on Titan, the generic Titan
customization block was split into separate pgi and intel blocks. This
lost support for pgicuda. A new pgicuda block was added. The location
of the Trilinos library (for pgi) on Titan was also fixed. No support
for Trilinos with intel, cray, or pgicuda is available at the moment
(though there is support for gnu).