changing index variable declarations from real to integer
Five "tracer indices" variables are declared real instead of
integer. The Cray compiler complains and aborts the build when
at least one of these are used as an array index. The declarations
are changed to integer. Note that this change is also in a more
recent version from NCAR, but as part of a much larger rewrite of
the ecosys routines.