terminating syslog monitoring script when ACME run completes
The syslog progress monitoring script is designed to run until
the batch job script kills it, which occurs after the ACME run completes. When
the job script is killed because requested time is exceeded, the syslog
job is not also terminated on Mira.

Here syslog is modified to test whether the remaining time is <= 0 or
if the ACME job is complete. If so, the job will then terminate itself. For
maintenance purposes the same modification is applied to all versions of syslog,
even though it has only been identified as needed on Mira.