wait_for_tests: Make sure cdash timestamp is in sync
Cdash was sometimes updating the TAG file based on the time
in the test XML file. This was causing the TAG file and the
dirname of the results to go out of sync, breaking the cdask
jenkins plugin.

I have run Jenkins on this branch and this commit fixed the
issue.

[BFB]

SEG-41