Fixed issues on Titan with craype-interlagos
The craype-interlagos module on Titan was changed to specify "-tp=bulldozer" in the compiler flags for PGI, as the omission of this was a performance bug, and the wrappers should be targetting the AMD interlagos architecture. However, there is a bug in PGI 14.2 and 14.10 ( possibly others ), which breaks reproducibility tests in ACME. I have unset the environment variable that adds this flag, so reproducibility is restored on Titan with this fix. I hope to find out what is leading to the issue, isolate the file(s) in Depends.pgi, and get things working with the correct architecture target soon. But in the interim, this allows the developer's test suite on Titan to pass.