Specifying a concrete default timelimit for all jobs on Cetus
A zero timelimit should ideally (as per qsub man pages) represent
the maximum allowed walltime for the job's queue. For the default
job queue on Cetus this timelimit is 1hr. However due to a bug
in qsub (and issues with ALCF job queue policy enforcement) we
 can no longer specify a 0 timelimit. So specifying a concrete
timelimit of 59min for now.

Also see Issue #137

SEG-81