Changes to XML files for easier use of RRM case
Three sets of changes in this commit.
1.   The namelist_defaults_cam.xml and the namelist_mod.F90 have updates
so that a user wanting to run RRM case does not have to specify the path
of the grid file in the user_nl_cam.  This will reduce potential error
and improve ease of the RRM capability.  It was found that the mesh_file
designation needed to be blank for a structured (non-RRM) grid, so
namelist_mod.F90 was altered to not have a model crash during runtime.

2.  The paramter EPS_AAREA needed to be raised in env_run.xml by a
factor of 10.  This was done by hand every time an RRM case was built.
The changes to config_grid.xml now have the correct value of EPS_AAREA
to run RRM.

3.  Final changes is to $PROJECT variable for SNL computers.

 On branch eroesler/ATM/RRM-meshfilefixer
 Changes to be committed:
	modified:   ../models/atm/cam/bld/namelist_files/namelist_defaults_cam.xml
	modified:   ../models/atm/cam/src/dynamics/se/share/namelist_mod.F90
	modified:   ccsm_utils/Case.template/config_grid.xml
	modified:   ccsm_utils/Machines/mkbatch.redsky
	modified:   ccsm_utils/Machines/mkbatch.skybridge