jenkins_generic_job: Don't process test results or wait_for_tests if --namelists-only
For some reason, when -nlcompareonly is passed to create_test, tests are
being left in the RUN state instead of PASS. This is causing Jenkins to report
namelist generation jobs as failing even though they are working fine. For now,
skip all test-result processing if we are only generating namelists.

[BFB]