acme_bisect: A new script to facilitate git bisections
Bisecting in acme would be non-trivial without a script due to the
complexity of running a test and the subtleties of interpreting test
results. We need a script to encapsulate these issues to make it very
simple to do bisections.

Also adds capability to wait_for_tests to not fail a test if its
only problem is a namelist diff. This will allow bisections to find
real test failures only.

[BFB]

SEG-102