Do not echo xmlchange commands to CaseStatus if run by CESM scripts
The purpose of echoing xmlchange commands to CaseStatus is to provide a "paper
trail" of changes made by the user, so calls to xmlchange by scripts that are
part of the normal case setup/build process should generally not be echoed. (The
"paper trail" should just contain things that the user would have to apply if
trying to recreate this case.)

I have added a -noecho option to xmlchange, and used that option in most of the
case setup/build scripts.

However, I am keeping the echoes in place for the test scripts (e.g., in
testcase_setup, and in individual tests), because these are cases where I think
it is useful to have a record of the changes made by the test scripts.

Also: fixed bug in cesm_build.pl: append to CaseStatus rather than overwriting
it.