jenkins_generic_job: Add new rebless mode
Prior to this change, blessing binary changes required the diffing
test to be re-run in -generate mode. It is much more efficient to just
copy the binary files from the most recent run of the test. This also
ensures that we know what we are blessing since there is no
opportunity for additional commits to be introduced between the time
of the diff and the time of the bless.

[BFB]

SEG-113