Moved precompute of hypervis to tracer prestage
Move a variable from advance_hypervis_scalar's prestage to the overall tracer rk2 prestage, where it can be overlapped with the DSS exchanges. Overall speedup of CUDA over default PGI is 2.7x for the gracers with 128 elements per node. This is a speed-up of 1.07x over the existing CUDA port before this branch.