Changing default walltime clock
When NANOTIMER is available, this is the timing library default.
NANOTIMER has potential problems with multicore processors, and this
modification changes the default to MPI_WTIME (when MPI_WTIME is
available). This has no impact on ACME currently as MPI_WTIME is
specified explicitly by the ACME scripts when running with MPI.