Fix unit test netCDF detection and NAG support.
This commit makes use of the FindNETCDF CMake module to detect the
netCDF library for unit tests. It also fixes some files that prevented
the tests from running with the NAG compiler.

This commit BREAKS support for NAG on goldbach, but that machine is on
its way out. When the successor machine comes online, a working NAG 6.0
port will be added.

Testing:

    Test suite: Unit tested only.
    Baseline: N/A
    Namelist changes: None
    Status: bit for bit

Code reviews:

    None yet.