Refactor to bless capability
* Pull bless functionality out of jenkins_generic_job and put into it's own tool
* Split compare_namelists and simple_compare into executables and libraries
  - Allows us to import the capability of these two programs instead of having to run from shell
* Add support for updating hist files in bless_test_results
* Have hist_compare.csh return non-zero error code on failure

Note: this usefulness of this tool is significantly reduced for restart tests
because, upon restart, many of the namelist files are re-created with different
values, so what's left in a testcase after the run is different than what would
be compared against baselines. Hopefully this unfortunate implementation will
be fixed someday.

[BFB]

SEG-113