get user defined machines in ~/.cime and systems without a batch system working
The perl batch system refactor broke the ability to run on systems without a batch system and the ability to create a user defined machine with configuration files in ~/.cesm. This restores that functionality.

Users can place custom config_*.xml files into HOME/.cime (previously HOME/.cesm) instead of modifying configuration files in cime/machines. The batch system generator will look at, in order: HOME/.cime/config_*.xml, CIMEROOT/machines/config_*.xml, and CASETOOLS/config_*.xml. If it finds a matching xml file, it searches for the specified machine or batch system in that file, and uses it. If it doesn't then it keeps searching until it finds the desired xml or dies if nothing is found. Specifying a batch system as 'none' now generates perl run scripts that do not worry about the batch directives. This does not update the functionality for compilers, mpi, etc that were not affected by the batch refactor.

There was a lot of copy and paste code in the Batch*.pm modules related to reading machine and batch xml files. They have been consolidated into reusable functions.

Tested with: ERP_D.f45_g37_rx1.A