Fix module cleanup in env_mach_specific.edison
The test scripts source the env_mach_specific file, and then
children processes re-source this same file for some testcases.
This means you need to be careful to avoid module conflicts and
enviroment problems.

In this case, the wrong modules were being cleaned up which left
some modules loaded to later conflict with other modules the env_mach
file was trying to load.

This commit fixes this issue.