Fix finding component history file for baseline gen / comp
Main diff is in component_compgen_baseline.sh:

-       test_hist=`cd $test_dir; ls -1
        *.${model}*.${extension}.*.nc.base 2>/dev/null | tail -1`
+       test_hist=`cd $test_dir; ls -1
        ${testcase}.${model}*.${extension}.*.nc.base 2>/dev/null | tail -1`

The original version was too greedy. This meant, for example, that in
a test that had 'cam' in its name (e.g., because of a cam testmods
directory), it would sometimes find a clm history file when it was
looking for a cam history file, because the clm history file name
matched the original version.

Changing this required passing the full testcase (i.e., CASE) into
component_compgen_baseline.sh.