Fixing ERS.f45_g37.B1850C5 on Cetus and Mira
This commit modifies the number of tasks and threads for the atm
FV 4-deg grid on Cetus and Mira so that the ERS.f45_g37.B1850C5
test passes on Cetus and Mira. Without the fix the test fails
with a "fv_decomp_set failed to find a decomposition" (from
CAM build-namelist). This test reduces the number of tasks
per component from the default, 512, to 8. All other machines
of interest to ACME use 8 tasks/component for the FV
4-deg grid.

This change only affects Cetus and Mira.

Issue #280

[BFB]