wait_for_tests: Increase wait interval in wait_for_tests
I just noticed there's a very sinister race condition in wait_for_tests.
There can be a breif period of time in the TestStatus file were all lines
are listed as pass before an additional line is added. If the wait_for_tests
thread happens to wake up during that period, the test will be recorded as PASS
before it has even completed. Also, if the test later fails, it won't be caught
by our dashboard.

This is a temporary solution until a more robust one can be found.

[BFB]