Merge branch 'rljacob/acme-merge2' (PR #329)
Bring the Common Infrastructure for Modeling the Earth (CIME) in to
ACME.

This brings in cime 2.0.14 as a subtree of
https://github.com/ACME-Climate/cime

Move the content of ACME/scripts to ACME/cime
Move the major component models from ACME/models to ACME/components.

Model run scripts are now in ACME/cime/scripts

Changes to the coupler result in non-bfb changes in answers.

Numerous other changes and improvements are summarized in the commits on this
branch.

[non-BFB]
SEG-110