Improvements to testing scripts
By default, all new test machines have been set to only test
acme_developer and to not maintain baselines. We can ramp-up
the rigor of the testing as the machine proves useful and low
latency. Edison testing is scaled-back due to incredible latency.

Jenkins driver can now specify whether to do baselines comparisons.

This work was inspired by the idea that maintaining baselines on 6
different machines for the same tests is somewhat redudant. The small
chance of a test diffing on one platform and not another is not
worth the cost. High latency machines make this even worse as the
feedback time on a diff is extremely slow. We basically just want
to make sure that ACME works on these machines for now.

[BFB]