Merge branch 'jnjohnsonlbl/scripts/update_inputdata_dir' into master (PR #326)
Switching ACME NERSC inputdata directory to acme project from ccsm1

For runs on NERSC machines, we've been borrowing input data and space
from the 'ccsm1' project, which is an account for work with CESM. We
have created our own inputdata directory for the acme project on
NERSC, and now need to populate that directory.

This commit switches us from the ccsm1 inputdata directory to the acme
one. The testing process will be affected until we populate this new
directory.

[BFB]

* jnjohnsonlbl/scripts/update_inputdata_dir:
  Switching ACME inputdata directory to acme project from ccsm1.