Added dynamic roots
This commit adds dynamic root functionality to ALM. Root distribution
is simulated as a weighted function of water and nitrogen availability.
When crops are active, a root depth function allows root growth from
the 2nd soil layer to a maximum depth for each crop type defined by
a new parameter in the parameter file. Root depth increases linearly
with growing degree days.

The module is turned on by default when use_vertsoilc is on. A namelist
parameter is also introduced, use_dynroot, that can activate dynamic roots.

A new parameter file for crops is defined with an additional
parameter, root_dmx, which is the maximum rooting depth for crops.

[Non-BFB]
[CC]

LG-56