OpenACC code working as beforehand
The OpenACC code is giving the same answer it was on the old HOMME branch. It differs from the new HOMME branch, and I believe this is likely because of a new implementation of the limiter_optim_iter_full, but I'm not sure yet. Next step is to ensure the same answer as the new HOMME on CPU.