OpenACC now machine precision same w/ CPU!
The OpenACC code is now machine precision the same with the CPU code after I change limiter_optim_iter_full. Now, I need to alter the viscosity_mod routines with min/max and edge_mod routines to match what is in the CPU code for optimization, but the answer at least is now the same.