Fixing PIO Source line too long issue
PIO build fails with the pgi compiler on blues during the nightly
builds with a "Source line too long" error message. This is
because of the expansion of the "__FILE__" macro to a long fully
qualified file name. This issue occurs on nightly builds even
with the "-Mextend" flag (allow source upto 132 columns).

Also note that the build will not fail (with pgi on blues) if
it is from a directory with a "not-too-long" fully qualified name.

Fixing this issue by using the custom "__PIO_FILE__" macro within
PIO that expands to only the current filename.

Fixes #455
[BFB]