jenkins_script: deinit submodules post-run
Jenkins randomly fails to fetch with populated submodules, even
with the 'do not process submodules' flag checked. This change will
force jenkins to deinitialize submodules at the end of each run which
should ensure that the subsequent fetch for the next night's run will
work. The initialization is happens in jenkins_generic_job.

This implementation won't work for timed-out jobs, but that hasn't
been a problem lately and we should be getting a better fix for
submodules on Jenkins soon.

[BFB]