create_test: Dodge 'bad interpreter' error in build phase
This error is only generated when the filesystem is modifying
the script as it's trying to be run. We dodge this error by
simply re-running until this error goes away.

Debugging by dumping all processes that are touching the test_build
script revealed that buildnml scripts are allegedly touching
test_build scripts in other cases. A visual inspection of the code
did not indicate any place were this might be happening.

One thing that is clear is that the buildnml scripts are wastefully
forking a huge number of processes to read/write XML. We could probably
greatly simply and speed up these scripts.

[BFB]