bless_test_results: Minor fixes and enhancments
Allow user to provide a test_id to bless_test_results. Melvin dumps
both master and next results into the same directory, so the test_id
is needed in order to ensure you just process one set of results.

Fixed a problem when asking for the CESMSCRATCH root for another user.
The old code only worked if the XML entry had $USER. If it used
ENV{HOME} instead, it didn't work. I added an additional replacement
to take this into account.

[BFB]