Add chemical mechanism for MAM4 with resuspension
A new chemical mechanism is needed for MAM4 with resuspension, since
the code adds new chemical species in the coarse mode aerosol.  This
update does the following:

- Adds the chemistry mechanism files (produced by campp) into the
  source code as the trop_mam4_resus package.
- Implements trop_mam4_resus in configure scripts.
- Adds cases to if-statements that depend on the chemistry package in
  use, where needed.

These changes are needed in order to be able to run the code
introduced in PR #416 without a user-provided chemical mechanism file.
The new configuration can be turned on by using the CAM_CONFIG_OPTS
options "-chem trop_mam4" and "-rain_evap_to_coarse_aero" in
combination.

[BFB]
[FCC]
AG-138