bless_test_results: simplify bless process
You should not need to pass the full test_id of the tests you want to
bless. When blessing from the nightlies, the compiler and branch name
should be sufficient. There are still some cases where it's only possible
to differentiate test results by test_id, so that option is kept but should
only rarely be needed.

[BFB]