Merge branch 'apcraig/mosart/add-mosart' (PR #526)
These changes add a new model, mosart, as a runoff model.
The changes to the scripts and coupling support those changes.
In addition to the new source code under mosart, there were also
small changes to rtm for consistency and a few changes to the coupling fields.

A handful of new compsets were created to run with mosart, but
additional compsets will be required as needed for science.

LG-20
[non-BFB]

Conflicts:
	cime/components/data_comps/datm/bld/namelist_files/namelist_definition_datm.xml
	cime/scripts/Tools/config_compsets.xml
	components/clm/bld/namelist_files/namelist_defaults_clm4_0.xml
	models/lnd/clm/src/cpl/lnd_import_export.F90
	scripts/ccsm_utils/Case.template/config_definition.xml