adding/changing timers to improve coverage in CAM/HOMME
With the update to the new HOMME branch, with the inclusion
of CLUBB and MG2, and with the new compsets, the timers
in the atmosphere are missing significant contributions to the
wallclock time. Based on recent experiments, timers have been
added or moved. This will likely have to be done again as the
V1 code continues to evolved, and when the OpenACC code for
CAM/HOMME is merged in, but this will help in the meantime.