Fix COSP lidar simulator floating invalid error (issue #621)
The denominator [EXP(-2.0*tautot_liq(:,k+1))] in the formula
calulating beta_perp_liq goes to zero when the value of
tautot_lay_liq is too large, leading to a 'divide by zero' error.
The same is for ice. A fix is given by constraining maximum value
of tautot_lay_liq(ice).

[BFB]