Change MPAS-O defaults for watercycle simulations
This commit updates the MPAS-O submodule and updates the namelist and
streams files. The namelist and streams changes are to setup the
defaults for the watercycle simulations.

Additionally this commit correctly depresses the ocean surface based on
the amount of sea ice above it, and corrects the computation of the melt
potential which is passed to the sea ice model.

Answers will be Non-BFB because of minor changes to namelist defaults
and the application of bug fixes.

[Non-BFB]