Change MPAS-CICE defaults for watercycle simulations
This commit updates the MPAS-CICE submodule and updates the namelist and
streams files. The namelist and streams changes are to setup the
defaults for the watercycle simulations.

Additionally, this commit fixes the conversion of the sea surface
temperature, and adds default time steps for the mpas120 and oEC60to30
meshes.

Answers will be Non-BFB because of minor changes to namelist defaults.

[Non-BFB]