Increased length of caseid variable
Some of the tests in the testing suites were failing beacuse of the
folowing error:

"CASEID must not exceed 79 characters"

This PR addresses this by increasing the length of variable "caseid"
so that casenames can be longer than 79 characters. A smoke test
with FC5 compset works fine with a long casename which otherwise
would crash with the error message shown above.

Fixes #689
[BFB]
SEG-211