create_test: Immediately submit to queue once build is done.
The old behavior was highly suboptimal in the case where there were more
builds to do than number of threads. This could result in a very long gap
in time between when a build was completed and when the testcase was
submitted to the queue.

This commit causes the run phase to be done as the last part of the build
phase if on a batch system.