Fixes build-namelist by excluding variables not required in RK MP
The code crashes in build-namelist if run using BM1850CN compset.
This compset uses RK microphysics (MP) for which micro_mg_dcs is not
defined. build-namelist script tries to add this variable to the
namelist and crashes as this variable is not defined for RK
microphyics. This PR moves this variable (along with ice_sed_ai)
to an "mg" microphysics if block so that they are not added to the
namelist when RK microphysics is selected.

Fixes #887
[BFB]