Cleanup ESMF output in cesm.log
Each MPI rank on Mira is printing messages of the form

cime/share/esmf_wrf_timemgr/ESMF_TimeMod.F90", line 840: 1525-076 A
comma was expected after the I edit descriptor in a format
specification.  The program will recover by assuming a comma.

This can extend to 50K such lines in a 68K line cesm.log file.
The issue is format specification
- (I0004.0004'-',I2.2,'-',I2.2,'_',I2.2,':',I2.2,':',I2.2)
which should be
- (I0004.0004,'-',I2.2,'-',I2.2,'_',I2.2,':',I2.2,':',I2.2)
to avoid the warning message.

[BFB] - Bit-For-Bit