Fix errors when running `cs.status.*`.
There were several errors related to baseline tests, including memory
and throughput baselines, when running the `cs.status.*` script. (This
is the script generated in the root directory for a test suite, for
checking the current status of all tests in the suite.)

Most of the necessary changes are in `system_tests_common`, where the
output produced using `append_status` needs to use a format and phase
identifiers that can be recognized by
`wait_for_tests.parse_test_status`. `parse_test_status` also needs to
recognize the memory and throughput comparisons as valid test phases, as
well as recognizing "BFAIL" (the status resulting from a *missing*
baseline directory) as a valid status.

Testing:

    There are no formal regression tests for this functionality, but the
    CAM tests on yellowstone have been run with these changes, and
    cs.status runs without throwing an exception.