Add white space after batch directives in script templates
This commit adds white space after the batch directives in the script
templates to support more batch queue systems. Moab, for example,
considers it a syntax error if there isn't a blank line after the batch
directives. Additionally, it ensures the batch direcitves always are at
the top of the script, as moab fails to accept the batch directives if
they are not.