Fix for model crashes on Mira when compiled with O3 optimization
The PR lowers the optimization to -O2 for two F90 files:
seasalt_model.F90
linoz_data.F90

in Depends.mira so that the model runs fine on Mira with O3 (default)
optimizations level.

This PR partially fixes issue #871. F-compsets with ne16 grid are still
crashing with a different runtime error.

[BFB]