Allow manage_case --query-machines to work and remove "(required)" from -mach
Remove "(required)" from documentation on "-mach" option to create_newcase
add note that when not given it gets it from hostname.

Get "./manage_case --query-machines" to work on a machine
that is not in the list of defined machines in the config_machines.xml
file.

Test suite:  None
Test baseline:  None
Test namelist changes:  None
Test status: bit for bit

./manage_case --query-machines works on a non-defined machine

and

./create_newcase --help works

Fixes: #479 #480

Code review: self,jedwards,rljacob