Read adiabatic or ideal_phys option from atmosphere namelist
Currently, the driver tells the atmosphere model if it is to use
an adiabatic or ideal physics (modified Held-Suarez) option.
However, this is not the driver's responsibility and this functionality
is going to be removed from CIME. This commit redirects the atmosphere
model to read this information from its own namelist.
Note that this commit should be merged before the CIME change is
imported. Note that this change is NOT in CIME 5.1.

[BFB]