Changes to handling of environment_variables
Incorporates jedwards' fix on another PR but in such a way
as to not break melvin. Specifically, we need to be able to use
environment variables set by the modules in environment_variables. Also,
if we're going to call get_resolved_value, then we cannot support bash
in this field