Better handling of timeouts for case.test
Add signal handlers for SIGTERM and SIGINT to ensure that a python
exception is generated when the case.test process is hit with these
signals. slurm documentation states that it will use a SIGTERM before
issuing a SIGKILL.

These changes will ensure that the RUN phase will be FAIL for timed-out
tests. Also, the last line in TestStatus.log will be 'test timed out'