Rename CMAKE_MODULE_DIRECTORY to CIME_CMAKE_MODULE_DIRECTORY
I like keeping with the CMake convention of prefixing variables that are
specific to our project with CIME (used to be CESM): This makes it clear
which variables are specific to our project, vs. which are internal
CMake variables.

Test suite: None
Test baseline: N/A
Test namelist changes: N/A
Test status: N/A

Fixes: None

User interface changes?: No

Code review: None