Directory renaming for model source code
As discussed in issues #1178 and #1150, the
four model source directories are moved to
a CIME/src directory. This is intended to help
people navigate cime more easily.

The changes are mainly in xml config:
  $CIMEROOT/externals -> $CIMEROOT/src/externals
repeated for driver_cpl, share, components.

Also, many changes in python to add src to the
relative directory path.

scripts_regression_tests all pass