clm4_5_14_r226
Update cime version and cism to version that similar to that in cesm2_0_beta05, but with many fixes in cime for
clm. Also update mosart to fix a bug on hobart. The update of cism does change answers as it makes CISM2 the
default with the new 4km greenland grid.

There are some user improvements in the latest cime version. Long term archiving issues are fixed. A file
called CASEROOT is added to your case's $RUNDIR so that you can easily see where your source code is. A few
options were added to create_newcase and create_clone (see details below).

NOTE: You will need to use the "--run-unsupported" to create_newcase for compsets/grids that aren't tested or supported!!!

Fix a few other existing bugs. Fix createcrop_landunit to work with the new CLM surface datasets. Bring in a
workaround for a compiler bug with intel 17 and 18. Bring in a version of PTCLM that works with the latest cime.