Big refactor of CaseStatus management
Write a special hanlder in utils.py in order to be able to
easily encapsulate the concept that a function needs to log
its statu to CaseStatus and be robust in the face of exceptions.

Rearrage and rename some functions in build.py in order to make
it more clear what is the public API and what are internal functions.