Fixed a bug in calculating the total frozen static energy, where the contribution of rain water was not considered.
We were aware of this before, but since it is a very small term (compared to Q and CLDLIQ) and we didn't want to break the BFB feature before the previous model freeze, this wasn't fixed at that time.

The change to the original total frozen static energy is small, but it will cause N-BFB results. The TOA net flux change compared to Qi's simulation on Edison is about 0.15W/m2 (years 2-6), but it is hard to tell how much of it is from the bug fix and how much is due to the machine difference (my simulation is on EOS).

Also removed some unnecessary comments.

[non-BFB]