bug fix for array-passing related crash of betr
With the most recent intel compiler, the betr code crashes when running on more than 24 cpus on edison.
The cause of this crash was traced and found related to the passing of arrays between different subroutines.
Now all arrays are passed with explicit specification of their lower and upper bounds, so betr can run without
noticable issue. However, the advection capability of betr has to be turned off, because the new intel compiler
does not work well with passing generic data types. A workaround is developed in betr-v2, and will be integrated
in the future.

This fix does not affect codes other than betr, so is bfb.