Get compiler from the Macros.cmake file
This works on roo2 with both serial and mpi. But note that on roo2, I
set the full path to SFC and MPIFC. It also works to rerun unit tests
out of an existing directory (using either serial or mpi) - only the
bare minimum is done.

(I was worried that it might not work to rebuild, based on this comment
in the Makefile:

)