Clean up documentation for check_map tool
Mostly making sure README is still accurate, but I added a little bit of
verbose output to area conservation check portion of the tool and also cleaned
up the Makefile so that loading the ESMF module is not required to clean the
build. There were also some files missing from .gitignore that make git status
a lot easier to parse after running the tool a few times.