Fix greather-then-roundoff diff in cos zen calc
For jday larger than one, the argument (2*pi*jday + lon) for cos() will be
reduced by the cos function, and as a result, some accuracy will be
lost.  The reduction is different for different machines and results in
greather-then-roundoff differences between machines for the same case.
Solution is to change argument to ((jday-floor(jday))*2*pi + lon)