Improve handling of hist comparisons in TestStatus.log
There were two issues. First, when a diff occurred, it generated a
monstrous amount of output in TestStatus.log. A more reasonable approach
is to handle it the same as a build or run failure: note the fail and point
the user to the relevant log file.

The other issue is users being unclear if a compare passed or failed. The pass
or fail is now explicitly printed to the log.