Update path to new Albany install: redsky, edison
A new MPAS Landice tag, to be pulled into ACME in
the next few weeks, will have an updated interface
to Albany/FELIX. I have started to upgrade the
Albany installation on targeted machines. In this
commit, we update the path in the ALBANY_PATH
variable in config_compiler.xml for some machines.

This makes that change for redsky/chama, edison,
and sandia-srn-sems machines.
ToDo: install new albany on Mira, Titan, Anvil,
and other machines.

The hope is that the test MPASLIALB in
acme_integration will pass with this change on
these machines when done in concert with
the MPAS Landice update.