clm4_5_16_r252
Bring in a new CLM50 parameter file that updates a few fields for prognostic crop with affect on crop albedo
(rholnir,taulnir,taulvis,xl). Bring in changes to fix urban albedo for nighttime from Keith Oleson. Rather than set
albedo's for the urban canyon all to 1.0, set them to the viewfactors, as that will result in combined albedo of
one, whereas the previous way ended up with albedo's greater than one.

Here's a description from Keith about the change in urban albedo. " For diffuse radiation, which is
independent of solar zenith angle, the road albedos appear to be exactly equal to the road-sky view
factor.  And the wall albedos are equal to the wall-sky view factor. So the albedos are the albedo
of the surface (1.0) times the appropriate view factor.  This makes sense.  All incident radiation
entering the canyon should be reflected to the sky and the contribution of each surface is according
to the view factor of each surface to the sky as determined by the height to width ratio.  The view
factors of the two walls and the road to the sky sums to 1.0.  So setting the diffuse albedos to
these values when the sun is below the horizon results in a scaled diffuse albedo of 1.0.  For direct
radiation, it's more complicated because the solar zenith angle is involved when running the
radiation model.  However, it seems reasonable to assume the same albedos for direct as for diffuse."