Truncate small h2osfc values to zero
This replaces the earlier code that only did this truncation in one
particular circumstance.

This seems mainly important to truncate small negative numbers to 0, but
it also seems like a good idea to truncate small positive numbers that
should have been 0.

For now I'm leaving in some error-checking code to verify that answers
change by no more than roundoff-level; I'll remove that in a later
commit.