Add more unit tests of system_tests_compare_two, fix multisubmit bug
The multisubmit bug related to the indentation of a block: this was
mistakenly being done for both phases of a multi-submit, whereas it only
should have been done after the second
phase. https://github.com/ESMCI/cime/pull/1830 had this correct, but the
indentation was incorrect in https://github.com/ESMCI/cime/pull/1837,
which is what came to master.