clm4_5_16_r258
Revert the changes to fire that came in clm4_5_16_r257. Add some notes that clarify that the code as is, is correct.
Set n_melt_glcmec=10.0 for clm5.0 (This should have happened in the fresh snow grain size change in clm4_5_16_r256).
Update mosart to a version that doesn't accumulate water in short rivers. This is done by explicitly making sure river length
isn't smaller than a minimum value.