Be more graceful when dealing with cprnc log.
We want to copy the cprnc log to the casedir if there's a DIFF, but there's
no reason to copy it if it's been copied before.

Also, just issue a warning, instead of failing, if this copy is not successful.

These changes make it much less likely that bless_test_results will fail due
to permissions issues.

[BFB]